refactor todos for searchParams
